member update/ file upload

+++ src/main/java/com/choong/spr/controller/MemberController.java
@@ -1,6 +1,9 @@
 package com.choong.spr.controller;
 
+import java.security.Principal;
 import java.util.List;
+
+import javax.servlet.http.HttpServletRequest;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -81,38 +84,58 @@ public class MemberController {
 	}
 	
 	@GetMapping("get")
-	public void getMember(String id, Model model) {
-		MemberDto member = service.getMemberById(id);
+	public String getMember(String id, Principal principal, HttpServletRequest request, Model model) {
 		
-		model.addAttribute("member", member);
+		if (hasAuthOrAdmin(id, principal, request)) {
+			MemberDto member = service.getMemberById(id);
+			model.addAttribute("member", member);
+
+			return null;
+		}
+
+		return "redirect:/member/login";
+	}
+	
+	private boolean hasAuthOrAdmin(String id, Principal principal, HttpServletRequest request) {
+		return request.isUserInRole("ROLE_ADMIN") || (principal != null && principal.getName().equals(id));
 	}
 	
 	@PostMapping("modify")
-	public String modifyMember(MemberDto dto, String oldPassword, RedirectAttributes rttr) {
-		boolean success = service.modifyMember(dto, oldPassword);
-		
-		if(success) {
-			rttr.addFlashAttribute("message", "회원 정보가 수정되었습니다.");
+	public String modifyMember(MemberDto member, String oldPassword, Principal principal, HttpServletRequest request, RedirectAttributes rttr) {
+				
+		if(hasAuthOrAdmin(member.getId(), principal, request)) {
+			boolean success = service.modifyMember(member, oldPassword);
+			
+			if(success) {
+				rttr.addFlashAttribute("message", "회원 정보가 수정되었습니다.");
+			} else {
+				rttr.addFlashAttribute("message", "회원 정보가 수정되지 않았습니다.");
+			}
+			
+			rttr.addFlashAttribute("member", member); // model object
+			rttr.addAttribute("id", member.getId()); // query string
+			
+			return "redirect:/member/get";
 		} else {
-			rttr.addFlashAttribute("message", "회원 정보가 수정되지 않았습니다.");
+			return "redirect:/member/login";			
 		}
-		
-		rttr.addFlashAttribute("member", dto); // model object
-		rttr.addAttribute("id", dto.getId()); // query string
-		
-		return "redirect:/member/get";
 	}
 	
 	@PostMapping("remove")
-	public String removeMember(MemberDto member, RedirectAttributes rttr) {
-		boolean success = service.removeMember(member);
+	public String removeMember(MemberDto member, Principal principal, HttpServletRequest request, RedirectAttributes rttr) {
 		
-		if(success) {
-			rttr.addFlashAttribute("message", "탈퇴되었습니다.");
-			return "redirect:/board/list";
+		if(hasAuthOrAdmin(member.getId(), principal, request)) {
+			boolean success = service.removeMember(member);
+			
+			if(success) {
+				rttr.addFlashAttribute("message", "탈퇴되었습니다.");
+				return "redirect:/board/list";
+			} else {
+				rttr.addAttribute("id", member.getId());
+				return "redirect:/member/get";
+			}
 		} else {
-			rttr.addAttribute("id", member.getId());
-			return "redirect:/member/get";
+			return "redirect:/member/login";
 		}
 	}
 	
@@ -120,4 +143,15 @@ public class MemberController {
 	public void loginPage() {
 		
 	}
+	
+	@GetMapping("initpw")
+	public void initpwPage() {
+		
+	}
+	@PostMapping("initpw")
+	public String initpwProcess(String id) {
+		service.initMemberPw(id);
+		return "redirect:/board/list";
+	}
+	
 }
